+++ src/app/(authenticated)/structures/[id]/(controles)/ControleAccordion.tsx
@@ -1,0 +1,25 @@
+import { ReactElement } from "react";
+import { AccordionTitle } from "@/app/components/AccordionTitle";
+import Accordion from "@codegouvfr/react-dsfr/Accordion";
+import styles from "@/app/components/common/Accordion.module.css";
+
+export const ControleAccordion = ({
+  title,
+  lastVisit,
+  children,
+}: Props): ReactElement => {
+  return (
+    <Accordion
+      label={<AccordionTitle title={title} lastVisit={lastVisit} />}
+      className={styles["custom-accordion"]}
+    >
+      {children}
+    </Accordion>
+  );
+};
+
+type Props = {
+  title: string;
+  lastVisit: Date;
+  children: ReactElement;
+};
