--- src/app/(authenticated)/structures/[id]/(header)/StructureHeader.module.css
@@ -1,6 +1,0 @@
-.container {
-  position: sticky;
-  top: 0;
-  background-color: var(--grey-1000-50);
-  z-index: 1;
-}
